+++ lib/phoenix/ui/components/button_group.ex
@@ -7,12 +7,10 @@ defmodule Phoenix.UI.Components.ButtonGroup do
   use Phoenix.UI, :component
 
   attr(:color, :string, default: "blue", values: Theme.colors())
-  attr(:disabled, :boolean, default: true)
   attr(:element, :string, default: "div")
   attr(:orientation, :string, default: "horizontal", values: ["horizontal", "vertical"])
   attr(:size, :string, default: "md", values: ["xs", "sm", "md", "lg"])
   attr(:square, :boolean, default: false)
-  attr(:type, :string, default: "button")
   attr(:variant, :string, default: "contained", values: ["contained", "icon", "outlined", "text"])
 
   slot(:button)
@@ -22,7 +20,7 @@ defmodule Phoenix.UI.Components.ButtonGroup do
 
   ## Examples
 
-      ```
+      ```heex
       <.button_group>
         <:button>
           One
@@ -36,8 +34,8 @@ defmodule Phoenix.UI.Components.ButtonGroup do
 
   """
   @spec button_group(Socket.assigns()) :: Rendered.t()
-  def button_group(prev_assigns) do
-    assigns = prev_assigns |> build_wrapper_attrs() |> normalize_buttons()
+  def button_group(assigns) do
+    assigns = assigns |> build_wrapper_attrs() |> normalize_buttons()
 
     ~H"""
     <.dynamic_tag {@wrapper_attrs}>