+++ lib/phoenix/ui/components/helper_text.ex
@@ -4,8 +4,10 @@ defmodule Phoenix.UI.Components.HelperText do
   """
   use Phoenix.UI, :component
 
-  attr(:disabled, :boolean, default: false)
-  attr(:element, :string, default: "div")
+  attr(:class, :string, doc: "Override the classes applied to the component.")
+  attr(:element, :string, default: "div", doc: "The HTML element to use, such as `div`.")
+  attr(:extend_class, :string, doc: "Extend existing classes applied to the component.")
+  attr(:rest, :global, doc: "Arbitrary HTML or phx attributes")
 
   slot(:inner_block, required: true)
 
@@ -14,36 +16,22 @@ defmodule Phoenix.UI.Components.HelperText do
 
   ## Examples
 
-      ```
       <.helper_text>
         We'll never share your email with anyone else.
       </.helper_text>
-      ```
 
   """
   @spec helper_text(Socket.assigns()) :: Rendered.t()
-  def helper_text(prev_assigns) do
-    assigns = build_helper_text_attrs(prev_assigns)
+  def helper_text(assigns) do
+    assigns = assign_class(assigns, ~w(
+      mt-2 text-sm text-slate-500 dark:text-slate-400 disabled:text-slate-400 dark:disabled:text-slate-500
+      invalid:hidden
+    ))
 
     ~H"""
-    <.dynamic_tag {@helper_text_attrs}>
+    <.dynamic_tag class={@class} name={@element} {@rest}>
       <%= render_slot(@inner_block) %>
     </.dynamic_tag>
     """
-  end
-
-  defp build_helper_text_attrs(assigns) do
-    class = build_class(~w(
-      mt-2 text-sm text-slate-500 dark:text-slate-400 disabled:text-slate-400 dark:disabled:text-slate-500
-      #{Map.get(assigns, :extend_class)}
-    ))
-
-    attrs =
-      assigns
-      |> assigns_to_attributes([:extend_class, :field, :form])
-      |> Keyword.put(:name, assigns[:element])
-      |> Keyword.put_new(:class, class)
-
-    assign(assigns, :helper_text_attrs, attrs)
   end
 end