+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.1.4 (2022-11-17)
+
+### Enhancements
+
+- Improves and adds error styles for text_field component.
+- Improves TextFilter live component
+- text_field related components
+  - form_group
+  - label
+  - error_tag
+  - helper_text
+- misc docs cleanup
+
 ## 0.1.3 (2022-11-15)
 
 ### Enhancements